Fixed #29. Added sparse About page. Added ability to make email public/private on user edit page. Added AchievementInfo model for #30.

+++ app/models/Follow.java
@@ -107,7 +107,9 @@ public class Follow extends Model {
 			return;
 		}
 		Follow follow = Follow.findByBoth(follower, followed);
-		follow.delete();
+		if (follow != null) {
+			follow.delete();
+		}
 	}
 	
 }
